# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mobile Nest BD storefront and order management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  productCount: number;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  icon: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  category: string;
  categorySlug: string;
  price: number;
  /** @nullable */
  compareAtPrice: number | null;
  image: string;
  description: string;
  stock: number;
  rating: number;
  reviewCount: number;
  featured: boolean;
  /** @nullable */
  badge: string | null;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  categorySlug: string;
  /** @minimum 0 */
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  /** @minLength 1 */
  image: string;
  description: string;
  /** @minimum 0 */
  stock: number;
  featured: boolean;
  /** @nullable */
  badge?: string | null;
}

export type ProductUpdate = ProductInput;

export interface CartItemInput {
  /** @minimum 1 */
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export type OrderInputPaymentMethod = typeof OrderInputPaymentMethod[keyof typeof OrderInputPaymentMethod];


export const OrderInputPaymentMethod = {
  cod: 'cod',
  bkash: 'bkash',
  nagad: 'nagad',
  rocket: 'rocket',
} as const;

export interface OrderInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 8 */
  mobile: string;
  /** @nullable */
  whatsapp?: string | null;
  /** @minLength 1 */
  district: string;
  /** @minLength 1 */
  upazila: string;
  /** @minLength 1 */
  postOffice: string;
  /** @minLength 1 */
  village: string;
  /** @minLength 1 */
  houseNo: string;
  /** @nullable */
  additionalAddress?: string | null;
  paymentMethod: OrderInputPaymentMethod;
  /** @minItems 1 */
  items: CartItemInput[];
}

export interface TrackOrderInput {
  /** @minLength 1 */
  orderId: string;
  /** @minLength 8 */
  mobile: string;
}

export interface OrderItem {
  productId: number;
  name: string;
  quantity: number;
  unitPrice: number;
  lineTotal: number;
  image: string;
}

export interface OrderConfirmation {
  orderId: string;
  status: string;
  subtotal: number;
  deliveryCharge: number;
  total: number;
  paymentStatus: string;
  items: OrderItem[];
}

export type TrackedOrder = OrderConfirmation & {
  customerName: string;
  createdAt: string;
  address: string;
};

export type AdminOrder = TrackedOrder & ({
  /** @nullable */
  whatsapp?: string | null;
  paymentMethod: string;
  paymentStatus: string;
});

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface AdminSummary {
  totalOrders: number;
  pendingOrders: number;
  deliveredOrders: number;
  totalSales: number;
  lowStockProducts: number;
  outOfStockProducts: number;
}

export interface Storefront {
  categories: Category[];
  featured: Product[];
  newArrivals: Product[];
  offers: Product[];
}

/**
 * Authentication required
 */
export type UnauthorizedResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type ListProductsParams = {
search?: string;
category?: string;
featured?: boolean;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};