# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mobile Nest BD storefront and order management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: storefront
    description: Customer storefront operations
  - name: orders
    description: Guest checkout and order tracking
  - name: admin
    description: Protected store administration operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /storefront:
    get:
      operationId: getStorefront
      tags: [storefront]
      summary: Get storefront overview
      responses:
        "200":
          description: Storefront overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Storefront"
  /categories:
    get:
      operationId: listCategories
      tags: [storefront]
      summary: List product categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [admin]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [admin]
      summary: Update a category
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteCategory
      tags: [admin]
      summary: Delete a category
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
  /products:
    get:
      operationId: listProducts
      tags: [storefront]
      summary: List and filter products
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: featured
          schema:
            type: boolean
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [storefront]
      summary: Get a product
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteProduct
      tags: [admin]
      summary: Delete a product
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a guest order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderConfirmation"
  /orders/track:
    post:
      operationId: trackOrder
      tags: [orders]
      summary: Track an order by order ID and mobile number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackOrderInput"
      responses:
        "200":
          description: Order tracking result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackedOrder"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get admin dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminOrder"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Change an order status
      parameters:
        - $ref: "#/components/parameters/OrderCode"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrder"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: integer
        minimum: 1
    OrderCode:
      in: path
      name: id
      required: true
      schema:
        type: string
        minLength: 1
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        productCount:
          type: integer
      required: [id, name, slug, icon, productCount]
    CategoryInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        icon:
          type: string
          minLength: 1
      required: [name, slug, icon]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        categorySlug:
          type: string
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        image:
          type: string
        description:
          type: string
        stock:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        featured:
          type: boolean
        badge:
          type: ["string", "null"]
      required: [id, name, slug, category, categorySlug, price, compareAtPrice, image, description, stock, rating, reviewCount, featured, badge]
    ProductInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        categorySlug:
          type: string
          minLength: 1
        price:
          type: number
          minimum: 0
        compareAtPrice:
          type: ["number", "null"]
        image:
          type: string
          minLength: 1
        description:
          type: string
        stock:
          type: integer
          minimum: 0
        featured:
          type: boolean
        badge:
          type: ["string", "null"]
      required: [name, slug, categorySlug, price, image, description, stock, featured]
    ProductUpdate:
      allOf:
        - $ref: "#/components/schemas/ProductInput"
    CartItemInput:
      type: object
      properties:
        productId:
          type: integer
          minimum: 1
        quantity:
          type: integer
          minimum: 1
      required: [productId, quantity]
    OrderInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        mobile:
          type: string
          minLength: 8
        whatsapp:
          type: ["string", "null"]
        district:
          type: string
          minLength: 1
        upazila:
          type: string
          minLength: 1
        postOffice:
          type: string
          minLength: 1
        village:
          type: string
          minLength: 1
        houseNo:
          type: string
          minLength: 1
        additionalAddress:
          type: ["string", "null"]
        paymentMethod:
          type: string
          enum: [cod, bkash, nagad, rocket]
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CartItemInput"
      required: [name, mobile, district, upazila, postOffice, village, houseNo, paymentMethod, items]
    TrackOrderInput:
      type: object
      properties:
        orderId:
          type: string
          minLength: 1
        mobile:
          type: string
          minLength: 8
      required: [orderId, mobile]
    OrderConfirmation:
      type: object
      properties:
        orderId:
          type: string
        status:
          type: string
        subtotal:
          type: number
        deliveryCharge:
          type: number
        total:
          type: number
        paymentStatus:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required: [orderId, status, subtotal, deliveryCharge, total, paymentStatus, items]
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        name:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        lineTotal:
          type: number
        image:
          type: string
      required: [productId, name, quantity, unitPrice, lineTotal, image]
    TrackedOrder:
      allOf:
        - $ref: "#/components/schemas/OrderConfirmation"
        - type: object
          properties:
            customerName:
              type: string
            createdAt:
              type: string
              format: date-time
            address:
              type: string
          required: [customerName, createdAt, address]
    AdminOrder:
      allOf:
        - $ref: "#/components/schemas/TrackedOrder"
        - type: object
          properties:
            whatsapp:
              type: ["string", "null"]
            paymentMethod:
              type: string
            paymentStatus:
              type: string
          required: [paymentMethod, paymentStatus]
    OrderStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
      required: [status]
    AdminSummary:
      type: object
      properties:
        totalOrders:
          type: integer
        pendingOrders:
          type: integer
        deliveredOrders:
          type: integer
        totalSales:
          type: number
        lowStockProducts:
          type: integer
        outOfStockProducts:
          type: integer
      required: [totalOrders, pendingOrders, deliveredOrders, totalSales, lowStockProducts, outOfStockProducts]
    Storefront:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        featured:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        newArrivals:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        offers:
          type: array
          items:
            $ref: "#/components/schemas/Product"
      required: [categories, featured, newArrivals, offers]