# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mobile Nest BD storefront and order management API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  icon: string;
}