# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mobile Nest BD storefront and order management API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}